+++ flowmachine/flowmachine/features/network/total_network_objects.py
@@ -11,7 +11,7 @@ at the network level.
 
 """
 
-from typing import List, Optional, Tuple
+from typing import List, Optional, Tuple, Union
 
 from ...core.context import get_db
 from ...core.mixins import GeoDataMixin
@@ -19,9 +19,9 @@ from ...core import location_joined_query, make_spatial_unit
 from ...core.spatial_unit import AnySpatialUnit
 from ...core.query import Query
 from ..utilities import EventsTablesUnion
-from flowmachine.utils import standardise_date
+from flowmachine.utils import standardise_date, Statistic
 
-valid_stats = {"avg", "max", "min", "median", "mode", "stddev", "variance"}
+
 valid_periods = ["second", "minute", "hour", "day", "month", "year"]
 
 
@@ -157,8 +157,8 @@ class AggregateNetworkObjects(GeoDataMixin, Query):
     ----------
     total_network_objects : TotalNetworkObjects
 
-    statistic : {'avg', 'max', 'min', 'median', 'mode', 'stddev', 'variance'}
-        Statistic to calculate, defaults to 'avg'.
+    statistic : Statistic
+        Statistic to calculate, defaults to Statistic.AVG
 
     aggregate_by : {'second', 'minute', 'hour', 'day', 'month', 'year', 'century'}
         A period definition to calculate statistics over, defaults to the one
@@ -178,17 +178,15 @@ class AggregateNetworkObjects(GeoDataMixin, Query):
 
     """
 
-    def __init__(self, *, total_network_objects, statistic="avg", aggregate_by=None):
+    def __init__(
+        self,
+        *,
+        total_network_objects,
+        statistic: Union[Statistic, str] = Statistic.AVG,
+        aggregate_by=None,
+    ):
         self.total_objs = total_network_objects
-        statistic = statistic.lower()
-        if statistic in valid_stats:
-            self.statistic = statistic
-        else:
-            raise ValueError(
-                "{} is not a valid statistic use one of {!r}".format(
-                    statistic, valid_stats
-                )
-            )
+        self.statistic = Statistic(statistic.lower())
         if aggregate_by is None:
             if self.total_objs.total_by == "second":
                 self.aggregate_by = "minute"
@@ -218,12 +216,9 @@ class AggregateNetworkObjects(GeoDataMixin, Query):
 
     def _make_query(self):
         group_cols = ",".join(self.spatial_unit.location_id_columns)
-        if self.statistic == "mode":
-            av_call = f"pg_catalog.mode() WITHIN GROUP(ORDER BY z.value)"
-        else:
-            av_call = f"{self.statistic}(z.value)"
+
         sql = f"""
-        SELECT {group_cols}, {av_call} as value,
+        SELECT {group_cols}, {self.statistic:z.value} as value,
         date_trunc('{self.aggregate_by}', z.datetime) as datetime FROM 
             ({self.total_objs.get_query()}) z
         GROUP BY {group_cols}, date_trunc('{self.aggregate_by}', z.datetime)