+++ flowmachine/flowmachine/features/raster/raster_statistics.py
@@ -68,9 +68,7 @@ class RasterStatistics(Query):
             band = [band]
 
         if statistic not in ["sum"]:
-            raise NotImplementedError(
-                "The statistic %s is not implemented." % statistic
-            )
+            raise NotImplementedError(f"The statistic {statistic} is not implemented.")
 
         if vector is not None and grouping_element is None:
             raise ValueError("Provide `grouping_element` alongside `vector`.")