+++ flowmachine/flowmachine/features/subscriber/contact_reference_locations_stats.py
@@ -7,10 +7,10 @@
 Statistics for the distance between subscriber's own modal
 location and its contacts' modal location.
 """
+from typing import Union
 
 from .metaclasses import SubscriberFeature
-
-valid_stats = {"count", "sum", "avg", "max", "min", "median", "stddev", "variance"}
+from ...utils import Statistic
 
 
 class ContactReferenceLocationStats(SubscriberFeature):
@@ -26,7 +26,7 @@ class ContactReferenceLocationStats(SubscriberFeature):
         A flowmachine Query instance that contains a subscriber column. In
         addition to that the query must have a spatial unit or the target
         geometry column that contains the subscribers' reference locations.
-    statistic : {'count', 'sum', 'avg', 'max', 'min', 'median', 'mode', 'stddev', 'variance'}, default 'sum'
+    statistic : Statistic, default Statistic.SUM
         Defaults to sum, aggregation statistic over the durations.
     geom_column:
         The column containing the subscribers' reference locations. This is
@@ -49,16 +49,14 @@ class ContactReferenceLocationStats(SubscriberFeature):
     """
 
     def __init__(
-        self, contact_balance, contact_locations, statistic="avg", geom_column=None
+        self,
+        contact_balance,
+        contact_locations,
+        statistic: Union[Statistic, str] = Statistic.SUM,
+        geom_column=None,
     ):
 
-        self.statistic = statistic.lower()
-        if self.statistic not in valid_stats:
-            raise ValueError(
-                "{} is not a valid statistic. Use one of {}".format(
-                    self.statistic, valid_stats
-                )
-            )
+        self.statistic = Statistic(statistic.lower())
 
         self.contact_locations_query = contact_locations
         self.contact_balance_query = contact_balance
@@ -110,7 +108,7 @@ class ContactReferenceLocationStats(SubscriberFeature):
                 ST_Distance(subscriber_geom_point::geography, msisdn_counterpart_geom_point::geography) / 1000 AS distance
             FROM (SELECT DISTINCT subscriber_geom_point, msisdn_counterpart_geom_point FROM L) L
         )
-        SELECT subscriber, {self.statistic}(distance) AS value
+        SELECT subscriber, {self.statistic:distance} AS value
         FROM (
             SELECT C.subscriber, C.msisdn_counterpart, D.distance
             FROM ({self.contact_balance_query.get_query()}) C, L, D