+++ flowmachine/flowmachine/features/subscriber/per_subscriber_aggregate.py
@@ -6,8 +6,7 @@
 
 from typing import List
 from flowmachine.features.subscriber.metaclasses import SubscriberFeature
-
-agg_methods = {"count", "sum", "avg", "max", "min", "median", "stddev", "variance"}
+from flowmachine.utils import Statistic
 
 
 class PerSubscriberAggregate(SubscriberFeature):
@@ -22,7 +21,7 @@ class PerSubscriberAggregate(SubscriberFeature):
             A query with a `subscriber` column
         agg_column: str
             The name of the column in `subscriber_query` to aggregate. Cannot be 'subscriber'.
-        agg_method: {"count", "sum", "avg", "max", "min", "median", "stddev", "variance"} default "avg"
+        agg_method: Statistic, default Statistic.AVG
             The method of aggregation to perform
 
         Examples
@@ -55,7 +54,7 @@ class PerSubscriberAggregate(SubscriberFeature):
         *,
         subscriber_query: SubscriberFeature,
         agg_column: str,
-        agg_method: str = "avg",
+        agg_method: Statistic = Statistic.AVG,
     ):
         if "subscriber" not in subscriber_query.column_names:
             raise ValueError("'subscriber' column not in subscriber_query")
@@ -63,23 +62,19 @@ class PerSubscriberAggregate(SubscriberFeature):
             raise ValueError(f"'{agg_column}' column not in subscriber_query")
         if agg_column == "subscriber":
             raise ValueError(f"'agg_column' cannot be 'subscriber'")
-        if agg_method not in agg_methods:
-            raise ValueError(
-                f"agg_method must be one of {agg_methods}, not '{agg_method}'"
-            )
 
         self.subscriber_query = subscriber_query
         self.agg_column = agg_column
-        self.agg_method = agg_method
+        self.agg_method = Statistic(agg_method.lower())
+        super(PerSubscriberAggregate, self).__init__()
 
     @property
     def column_names(self) -> List[str]:
         return ["subscriber", "value"]
 
     def _make_query(self):
-        sql = f"""
-SELECT subscriber, {self.agg_method}({self.agg_column}) AS value
-FROM ({self.subscriber_query.get_query()}) AS sub_table
-GROUP BY subscriber
-"""
-        return sql
+        return f"""
+            SELECT subscriber, {self.agg_method:{self.agg_column}} AS value
+            FROM ({self.subscriber_query.get_query()}) AS sub_table
+            GROUP BY subscriber
+            """