+++ flowmachine/flowmachine/features/subscriber/subscriber_stay_lengths.py
@@ -7,8 +7,7 @@ from typing import List
 from flowmachine.features.subscriber.metaclasses import SubscriberFeature
 from flowmachine.features.utilities.subscriber_locations import BaseLocation
 from flowmachine.core.errors import InvalidSpatialUnitError
-
-valid_stats = {"count", "sum", "avg", "max", "min", "median", "stddev", "variance"}
+from flowmachine.utils import Statistic
 
 
 class SubscriberStayLengths(SubscriberFeature):
@@ -27,21 +26,20 @@ class SubscriberStayLengths(SubscriberFeature):
         per subscriber (or NULL location for subscribers that are active but
         unlocatable). The list is assumed to be sorted into ascending
         chronological order.
-    statistic :  {'count', 'sum', 'avg', 'max', 'min', 'median', 'stddev', 'variance'}, default 'max'
+    statistic :  Statistic, default Statistic.MAX
         Aggregation statistic over the stay lengths. Defaults to max.
     """
 
-    def __init__(self, *, locations: List[BaseLocation], statistic: str = "max"):
+    def __init__(
+        self, *, locations: List[BaseLocation], statistic: Statistic = Statistic.MAX
+    ):
         self.locations = locations
         if len(set(l.spatial_unit for l in self.locations)) > 1:
             raise InvalidSpatialUnitError(
                 "SubscriberStayLengths requires all input locations to have the same spatial unit"
             )
-        self.statistic = statistic.lower()
-        if self.statistic not in valid_stats:
-            raise ValueError(
-                f"'{self.statistic}' is not a valid statistic. Use one of {valid_stats}"
-            )
+        self.statistic = Statistic(statistic.lower())
+
         super().__init__()
 
     @property
@@ -60,7 +58,7 @@ class SubscriberStayLengths(SubscriberFeature):
 
         # Find stay lengths using gaps-and-islands approach
         sql = f"""
-        SELECT subscriber, {self.statistic}(stay_length) AS value
+        SELECT subscriber, {self.statistic:stay_length} AS value
         FROM (
             SELECT subscriber, count(*) AS stay_length
             FROM (